関係性に合わせ、DirectoryEntryをLs::Fileに変更
