+++ 08.ls_object/bin/ls.rb
@@ -1,9 +1,9 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
-require_relative '../lib/ls_formatter'
+require_relative '../lib/ls/formatter'
 
 if __FILE__ == $PROGRAM_NAME
-  ls = LsFormatter.new
+  ls = Ls::Formatter.new
   puts ls.output
 end
